@@ -4,7 +4,7 @@ version: 1.0.6
 homepage: https://github.com/igrik12/battery_info.git
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
   flutter: '>=1.20.0 <2.0.0'
 
 dependencies:
